Correct precedence for memoization in Auth::SSL
The #if takes precedence over ||=, resulting in values set explicitly
via #cert= and #cert_key= being ignored unless the #cert_key_file
and #cert_file, respectively, are set.

Add parens so that the #if determines whether to execute File#read
instead of blocking the implicit return of @cert and #cert_key.