Fixed where the "Cookie" header can be invalid
When there are no cookies, the "Cookies" key shouldn't exist.  The
existing behaviour was to allow it to be nil.

initialize_http_header in module HTTPHeader (net/http.rb) iterates over
all headers.  When "Cookie" is set and is nil, this method freaks out
when trying to call value.strip.  If the "Cookie" header has a value of
nil, it should not be included in the headers hash.